Convert mem_spec_limit, free_mem, and other memory formats to uint64_t.

+++ src/common/job_resources.h
@@ -99,24 +99,24 @@
  * updated (e.g. cpus and mem_used on that node cleared).
  */
 struct job_resources {
-	bitstr_t *	core_bitmap;
-	bitstr_t *	core_bitmap_used;
-	uint32_t	cpu_array_cnt;
-	uint16_t *	cpu_array_value;
-	uint32_t *	cpu_array_reps;
-	uint16_t *	cpus;
-	uint16_t *	cpus_used;
-	uint16_t *	cores_per_socket;
-	uint32_t *	memory_allocated;
-	uint32_t *	memory_used;
-	uint32_t	nhosts;
-	bitstr_t *	node_bitmap;
-	uint32_t	node_req;
-	char *		nodes;
-	uint32_t	ncpus;
-	uint32_t *	sock_core_rep_count;
-	uint16_t *	sockets_per_node;
-	uint8_t		whole_node;
+	bitstr_t *core_bitmap;
+	bitstr_t *core_bitmap_used;
+	uint32_t  cpu_array_cnt;
+	uint16_t *cpu_array_value;
+	uint32_t *cpu_array_reps;
+	uint16_t *cpus;
+	uint16_t *cpus_used;
+	uint16_t *cores_per_socket;
+	uint64_t *memory_allocated;
+	uint64_t *memory_used;
+	uint32_t  nhosts;
+	bitstr_t *node_bitmap;
+	uint32_t  node_req;
+	char	 *nodes;
+	uint32_t  ncpus;
+	uint32_t *sock_core_rep_count;
+	uint16_t *sockets_per_node;
+	uint8_t   whole_node;
 };
 
 /*
