+++ src/slurmd/slurmd/slurmd.h
@@ -80,7 +80,7 @@ typedef struct slurmd_config {
 	uint16_t     threads;           /* thread per core count           */
 	char         *cpu_spec_list;    /* cpu specialization list         */
 	uint16_t     core_spec_cnt;     /* core specialization count       */
-	uint32_t     mem_spec_limit;    /* memory specialization limit     */
+	uint64_t     mem_spec_limit;    /* memory specialization limit     */
 	uint16_t     cores;             /* core per socket  count          */
 	uint16_t     conf_cpus;         /* conf file logical processors    */
 	uint16_t     conf_boards;       /* conf file boards count          */
